refactor: restructure compact mode label
